Solving problems in sock shop kubernetes yaml
The sock shop application had some problems: the HOST/cart/* uri in the
ingress gateway (used by the front-end) was defined as carts, therefore
the front end was recieving 404's. After that, it was discovered
that the mongodb version used by carts-db deprecated a feature, so a
previous version (still lts) was used instead.

There is still a problems in the customer-orders front end path that is
still being worked on to make the microservices demo fully functional.